# api/openapi.yaml
openapi: 3.0.3
info:
  title: ForgeC2 Command & Control API
  description: |
    ForgeC2 is a professional-grade command and control framework for red team operations.

    ## Authentication
    Most endpoints require a session cookie (`forgec2_session`) from `POST /login`.
    Unauthenticated access to `/api/*` returns **JSON 401** (`{success:false,error:...}`), not HTML.

    ## CSRF
    State-changing requests need double-submit CSRF: cookie `forgec2_csrf` + header `X-CSRF-Token`.

    ## Spec quality (honest)
    - Route inventory is kept near-complete (CI coverage ≥90%, stale=0).
    - **Core operator paths** (health, agents, modules, listeners, dashboard, deploy)
      include request/response schemas; many secondary routes remain summary-only.
    - Dual-use paths (e.g. `GET /agents`) serve SPA HTML or JSON depending on Accept/auth.
    - `/debug/pprof/*` is intentionally omitted (disabled by default).

    ## Rate Limiting
    - Login: Limited attempts with lockout
    - API: Token bucket rate limiting per user

    ## Permissions
    Role-based access control (custom roles supported):
    - **admin**: Full control (all permissions)
    - **user**: Standard operator permissions (agents, tasks, credentials, limited settings)

  version: 2.4.1
  contact:
    name: ForgeC2 Support
    email: support@forgec2.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
...
  - url: https://c2.example.com
...
paths:
...
  /api/agents:
    get:
      tags: [Agents]
      summary: List all agents
      description: Retrieve a paginated list of all agents with optional filters
      operationId: listAgents
      security:
        - SessionCookie: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by hostname, username, or IP
        - name: status
          in: query
          schema:
            type: string
            enum: [online, stale, offline]
          description: Filter by agent status
        - name: os
          in: query
          schema:
            type: string
          description: Filter by operating system
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
          description: Items per page
        - name: sort_key
          in: query
          schema:
            type: string
            enum: [hostname, username, os, ip, last_seen, version, status]
            default: last_seen
          description: Sort by column (status orders by derived online/stale/offline)
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort direction
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /extc2/channels:
    get:
      tags: [Integrations]
      summary: External C2 channels
      operationId: extc2Channels
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Channels
...
  /api/v1/agents:
    get:
      tags: [Agents]
      summary: List agents (REST v1)
      operationId: apiV1ListAgents
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /agents:
    get:
      tags: [Agents]
      summary: List agents (dual-use SPA/API path)
      description: |
        With `Accept: application/json` returns agent list JSON.
        Browser navigation may receive SPA HTML. Unauthenticated JSON requests return 401.
      operationId: listAgentsLegacy
      security:
        - SessionCookie: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [online, stale, offline]
      responses:
        '200':
          description: Agent list (JSON) or SPA shell
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: forgec2_session
      description: Session cookie obtained from login endpoint
  schemas:
...
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
        details:
          type: string
...
    AgentListResponse:
      type: object
      description: |
        Envelope varies by handler. Common JSON forms include
        `{ agents: [...] }` or paginated `{ Agents|agents, Total, Page }`.
      properties:
        success:
          type: boolean
        agents:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique agent identifier
        hostname:
          type: string
        username:
          type: string
        os:
          type: string
          example: Windows 10 Pro
        arch:
          type: string
          example: amd64
        ip:
          type: string
        public_ip:
          type: string
        country:
          type: string
        city:
          type: string
        last_seen:
          type: string
          format: date-time
        status:
          type: string
          enum: [online, stale, offline]
        notes:
          type: string
        tags:
          type: string
        listener_id:
          type: integer
        parent_id:
          type: string
          description: Parent agent ID for P2P chaining
        p2p_mode:
          type: string
          enum: [smb, tcp, ""]
        p2p_listen_addr:
          type: string
        version:
          type: string
        pid:
          type: integer
        process_name:
          type: string
        integrity:
          type: string
          enum: [Low, Medium, High, System]
        elevated:
          type: boolean
        domain:
          type: string
        current_interval:
          type: integer
          description: Sleep interval in seconds
        current_jitter:
          type: integer
          description: Jitter percentage
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
...
    Unauthorized:
      description: Unauthorized - not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...